Testing author name with some random source code change
To manage repo-wise authors, migrate to the root of the repo:
- git config user.name "Your user name"
- git config user.email user@example.com

To have a default fallback, set up a global author:
- git config --global user.name "Your global user name"
- git config --global user.email global_user@example.com

If specified, local user configuration precede the global ones.

To reset author for multiple commits:
- git rebase -i <BASE_SHA> -x "git commit --amend --reset-athor"
This shall change author of all previous commits to the one
specified in the repo's local git config, or else the global config.